[refactor] adjust shared related import path
