+++ eLibrary/asgi.py
@@ -1,5 +1,5 @@
 from channels.routing import ProtocolTypeRouter, URLRouter
-from shared.channel_jwt_validate import JWTAuthMiddlewareStack
+from shared.validations import JWTAuthMiddlewareStack
 from chat import routing
